% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscOrganisation.R
\name{calcInverse}
\alias{calcInverse}
\title{This function is used to calculate the inverse of a matrix using cholesky decomposition}
\usage{
calcInverse(mat)
}
\description{
This function is used to calculate the inverse of a matrix using cholesky decomposition
}
\examples{
cat_function()
}
