% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpRegression.R
\name{covarianceMatrix}
\alias{covarianceMatrix}
\title{This program is to use to calculate the covariance matrix for given values for a given kernel}
\usage{
covarianceMatrix(params, X1, X2, S, kernel)
}
\description{
This program is to use to calculate the covariance matrix for given values for a given kernel
}
\examples{
cat_function() 
}
